% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestFitM2.R
\name{bestFitM2}
\alias{bestFitM2}
\title{Best fitting model2}
\usage{
bestFitM2(data, x = "x", y = "y")
}
\arguments{
\item{data}{a dataframe containing the variables x and y.}

\item{x}{name of the independent variable x.}

\item{y}{name of the dependent variable y.}
}
\value{
a list with 2 elements:
\item{BestFitM2}{the best fitting model}
\item{AllModel}{all seven basic models, NA represents the model is not applicable.}
}
\description{
\code{bestFitM2} selected the best fitting model from five. 
basic fitting models, according to AIC and BIC values.
}
\details{
This function selected the best fitting model from five 
basic fitting models.The five basic models include 
one simple linear fitting model(line2P), one quadratic curve model(line3P),
one logarithmic model(log2P), one exponential models(exp2P), and one power-law models(power2P).
This function is selected, when "bestFitM" reported some error messages.
}
\examples{
 
data("mtcars")
bestFitM2(data= mtcars, x= "mpg", y = "disp")
}
\author{
Ruilin Huang <fhruilin@163.com>
}
