% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitM.R
\name{FitM}
\alias{FitM}
\title{Fitting model}
\usage{
FitM(
  data,
  x = "x",
  y = "y",
  model = "line2P",
  summary = TRUE,
  Pvalue.corrected = TRUE
)
}
\arguments{
\item{data}{a dataframe containing the variables x and y.}

\item{x}{name of the independent variable x.}

\item{y}{name of the dependent variable y.}

\item{model}{the best fitting model.}

\item{summary}{summarizing the model fits. Default is TRUE.}

\item{Pvalue.corrected}{if P-value corrected or not, the value is one of c("TRUE", "FALSE").}
}
\value{

}
\description{
\code{FitM} calculate the parameters of the selected model.
}
\details{
This function calculate the parameters of the selected model. 
There are seven model parameters that can be calculated using this function.
The seven basic models include one simple linear fitting model(line2P),
one quadratic curve model(line3P),
one logarithmic model(log2P), 
two exponential models(exp2P and exp3P), 
and two power-law models(power2P and power3P).
}
\examples{
data("mtcars")
FitM(data= mtcars, x= "mpg", y = "disp",model = "line2P")
}
\author{
Ruilin Huang <fhruilin@163.com>
}
